--- mesa/visualization/templates/css/index_new.css
@@ -1,20 +1,0 @@
-:root {
-  --nav-height: 4.375rem;
-  --sidebar-width: 300px;
-  --primary-hue: 229;
-  --nav-bg: hsl(229, 6%, 83%);
-}
-
-aside {
-  margin-bottom: var(--nav-height);
-  width: var(--sidebar-width);
-  position: fixed;
-  margin-right: 8px;
-}
-
-main {
-  margin-top: var(--nav-height);
-  margin-bottom: var(--nav-height);
-  margin-left: var(--sidebar-width);
-  padding: 8px;
-}
