+++ mesa/visualization/templates/css/visualization.css
@@ -1,19 +1,23 @@
-.model-parameter {
-    margin-bottom: 15px;
+:root {
+    --nav-height: 4.375rem;
+    --nav-bg: hsl(229, 6%, 83%);
 }
 
-div.tooltip {
-    position: absolute;
-    text-align: center;
-    padding: 1px;
-    font: 20px sans-serif;
-    background: lightsteelblue;
-    border: 3px;
-    border-radius: 8px;
-    pointer-events: none;
+.container {
+    margin: var(--nav-height) auto;
+    max-width: 800px;
+    display: flex;
 }
 
-canvas.world-grid {
-    position:absolute;
-    border:1px dotted
+aside {
+    padding: 8px;
+    width: 250px;
+}
+
+input-parameters {
+    position: fixed;
+}
+
+main {
+    padding: 8px;
 }
